requirements.txt
google-cloud-pubsub==2.33.0
python-dateutil==2.9.0.post0
google-cloud-bigquery==3.38.0